move UnitTimes integration tests to separate module
